+++ src/sqlfluff/core/rules/base.py
@@ -404,7 +404,9 @@ class BaseRule:
     @classmethod
     def make_whitespace(cls, raw, pos_marker):
         """Make a whitespace segment."""
-        WhitespaceSegment = RawSegment.make(" ", name="whitespace", type="whitespace")
+        WhitespaceSegment = RawSegment.make(
+            " ", name="whitespace", type="whitespace", is_whitespace=True
+        )
         return WhitespaceSegment(raw=raw, pos_marker=pos_marker)
 
     @classmethod