+++ src/sqlfluff/dialects/dialect_ansi.py
@@ -52,7 +52,7 @@ ansi_dialect = Dialect("ansi", root_segment_name="FileSegment")
 ansi_dialect.set_lexer_struct(
     [
         # name, type, pattern, kwargs
-        ("whitespace", "regex", r"[\t ]+", dict(type="whitespace")),
+        ("whitespace", "regex", r"[\t ]+", dict(type="whitespace", is_whitespace=True)),
         (
             "inline_comment",
             "regex",
@@ -66,9 +66,14 @@ ansi_dialect.set_lexer_struct(
             dict(
                 is_comment=True,
                 type="comment",
-                subdivide=dict(type="newline", name="newline", regex=r"\r\n|\n"),
+                subdivide=dict(
+                    type="newline", name="newline", regex=r"\r\n|\n", is_whitespace=True
+                ),
                 trim_post_subdivide=dict(
-                    type="whitespace", name="whitespace", regex=r"[\t ]+"
+                    type="whitespace",
+                    name="whitespace",
+                    regex=r"[\t ]+",
+                    is_whitespace=True,
                 ),
             ),
         ),
@@ -85,7 +90,7 @@ ansi_dialect.set_lexer_struct(
         ("not_equal", "regex", r"!=|<>", dict(is_code=True)),
         ("greater_than_or_equal", "regex", r">=", dict(is_code=True)),
         ("less_than_or_equal", "regex", r"<=", dict(is_code=True)),
-        ("newline", "regex", r"\r\n|\n", dict(type="newline")),
+        ("newline", "regex", r"\r\n|\n", dict(type="newline", is_whitespace=True)),
         ("casting_operator", "regex", r"::", dict(is_code=True)),
         ("concat_operator", "regex", r"\|\|", dict(is_code=True)),
         ("equals", "singleton", "=", dict(is_code=True)),
@@ -786,11 +791,8 @@ class PartitionClauseSegment(BaseSegment):
         "PARTITION",
         "BY",
         Indent,
-        OneOf(
-            # Brackets are optional in a partition by statement
-            Bracketed(Delimited(Ref("ExpressionSegment"))),
-            Delimited(Ref("ExpressionSegment")),
-        ),
+        # Brackets are optional in a partition by statement
+        OptionallyBracketed(Delimited(Ref("ExpressionSegment"))),
         Dedent,
     )
 
@@ -816,24 +818,15 @@ ansi_dialect.add(
 
 
 @ansi_dialect.segment()
-class TableExpressionSegment(BaseSegment):
+class FromExpressionElementSegment(BaseSegment):
     """A table expression."""
 
-    type = "table_expression"
+    type = "from_expression_element"
     match_grammar = Sequence(
-        Indent,
         Ref("PreTableFunctionKeywordsGrammar", optional=True),
-        OneOf(
-            Ref("MainTableExpressionSegment"),
-            Bracketed(Ref("MainTableExpressionSegment")),
-        ),
+        OptionallyBracketed(Ref("TableExpressionSegment")),
         Ref("AliasExpressionSegment", optional=True),
         Ref("PostTableExpressionGrammar", optional=True),
-        Dedent.when(indented_joins=False),
-        AnyNumberOf(
-            Ref("JoinClauseSegment"), Ref("JoinLikeClauseGrammar"), optional=True
-        ),
-        Dedent.when(indented_joins=True),
     )
 
     def get_eventual_alias(self) -> Optional[AliasInfo]:
@@ -846,7 +839,7 @@ class TableExpressionSegment(BaseSegment):
 
         """
         alias_expression = self.get_child("alias_expression")
-        ref = self.get_child("main_table_expression").get_child("object_reference")
+        ref = self.get_child("table_expression").get_child("object_reference")
         if alias_expression:
             # If it has an alias, return that
             segment = alias_expression.get_child("identifier")
@@ -866,10 +859,30 @@ class TableExpressionSegment(BaseSegment):
 
 
 @ansi_dialect.segment()
-class MainTableExpressionSegment(BaseSegment):
+class FromExpressionSegment(BaseSegment):
+    """A from expression segment."""
+
+    type = "from_expression"
+    match_grammar = Sequence(
+        Indent,
+        OneOf(
+            # check first for MLTableExpression, because of possible FunctionSegment in MainTableExpression
+            Ref("MLTableExpressionSegment"),
+            Ref("FromExpressionElementSegment"),
+        ),
+        Dedent.when(indented_joins=False),
+        AnyNumberOf(
+            Ref("JoinClauseSegment"), Ref("JoinLikeClauseGrammar"), optional=True
+        ),
+        Dedent.when(indented_joins=True),
+    )
+
+
+@ansi_dialect.segment()
+class TableExpressionSegment(BaseSegment):
     """The main table expression e.g. within a FROM clause."""
 
-    type = "main_table_expression"
+    type = "table_expression"
     match_grammar = OneOf(
         Ref("BareFunctionSegment"),
         Ref("FunctionSegment"),
@@ -927,10 +940,10 @@ class WildcardExpressionSegment(BaseSegment):
 
 
 @ansi_dialect.segment()
-class SelectTargetElementSegment(BaseSegment):
+class SelectClauseElementSegment(BaseSegment):
     """An element in the targets of a select statement."""
 
-    type = "select_target_element"
+    type = "select_clause_element"
     # Important to split elements before parsing, otherwise debugging is really hard.
     match_grammar = GreedyUntil(
         "FROM",
@@ -975,7 +988,11 @@ class SelectClauseSegment(BaseSegment):
     type = "select_clause"
     match_grammar = StartsWith(
         Sequence("SELECT", Ref("WildcardExpressionSegment", optional=True)),
-        terminator=OneOf("FROM", "LIMIT", Ref("SetOperatorSegment")),
+        terminator=OneOf(
+            "FROM",
+            "LIMIT",
+            Ref("SetOperatorSegment"),
+        ),
         enforce_whitespace_preceeding_terminator=True,
     )
 
@@ -984,7 +1001,7 @@ class SelectClauseSegment(BaseSegment):
         Ref("SelectClauseModifierSegment", optional=True),
         Indent,
         Delimited(
-            Ref("SelectTargetElementSegment"),
+            Ref("SelectClauseElementSegment"),
             allow_trailing=True,
         ),
         # NB: The Dedent for the indent above lives in the
@@ -1018,7 +1035,7 @@ class JoinClauseSegment(BaseSegment):
         "JOIN",
         Indent,
         Sequence(
-            Ref("TableExpressionSegment"),
+            Ref("FromExpressionElementSegment"),
             # NB: this is optional
             OneOf(
                 # ON clause
@@ -1051,8 +1068,8 @@ class JoinClauseSegment(BaseSegment):
 
     def get_eventual_alias(self) -> AliasInfo:
         """Return the eventual table name referred to by this join clause."""
-        table_expression = self.get_child("table_expression")
-        return table_expression.get_eventual_alias()
+        from_expression_element = self.get_child("from_expression_element")
+        return from_expression_element.get_eventual_alias()
 
 
 @ansi_dialect.segment()
@@ -1076,7 +1093,18 @@ ansi_dialect.add(
 
 @ansi_dialect.segment()
 class FromClauseSegment(BaseSegment):
-    """A `FROM` clause like in `SELECT`."""
+    """A `FROM` clause like in `SELECT`.
+
+    NOTE: thiis is a delimited set of table expressions, with a variable
+    number of optional join clauses with those table expressions. The
+    delmited aspect is the higher of the two such that the following is
+    valid (albeit unusual):
+
+    ```
+    SELECT *
+    FROM a JOIN b, c JOIN d
+    ```
+    """
 
     type = "from_clause"
     match_grammar = StartsWith(
@@ -1087,11 +1115,7 @@ class FromClauseSegment(BaseSegment):
     parse_grammar = Sequence(
         "FROM",
         Delimited(
-            OneOf(
-                # check first for MLTableExpression, because of possible FunctionSegment in MainTableExpression
-                Ref("MLTableExpressionSegment"),
-                Ref("TableExpressionSegment"),
-            ),
+            Ref("FromExpressionSegment"),
         ),
     )
 
@@ -1101,12 +1125,15 @@ class FromClauseSegment(BaseSegment):
         Comes as a list of tuples (table expr, tuple (string, segment, bool)).
         """
         buff = []
-        direct_table_children = self.get_children("table_expression")
-        join_clauses = [
-            join_clause
-            for child in direct_table_children
-            for join_clause in child.get_children("join_clause")
-        ]
+        direct_table_children = []
+        join_clauses = []
+
+        for from_expression in self.get_children("from_expression"):
+            direct_table_children += from_expression.get_children(
+                "from_expression_element"
+            )
+            join_clauses += from_expression.get_children("join_clause")
+
         # Iterate through the potential sources of aliases
         for clause in (*direct_table_children, *join_clauses):
             ref: AliasInfo = clause.get_eventual_alias()
@@ -1115,7 +1142,7 @@ class FromClauseSegment(BaseSegment):
             table_expr = (
                 clause
                 if clause in direct_table_children
-                else clause.get_child("table_expression")
+                else clause.get_child("from_expression_element")
             )
             if ref:
                 buff.append((table_expr, ref))
@@ -1344,13 +1371,7 @@ class WhereClauseSegment(BaseSegment):
     parse_grammar = Sequence(
         "WHERE",
         Indent,
-        OneOf(
-            Bracketed(
-                # expression could be in brackets
-                Ref("ExpressionSegment"),
-            ),
-            Ref("ExpressionSegment"),
-        ),
+        OptionallyBracketed(Ref("ExpressionSegment")),
         Dedent,
     )
 
@@ -1438,12 +1459,7 @@ class HavingClauseSegment(BaseSegment):
     parse_grammar = Sequence(
         "HAVING",
         Indent,
-        OneOf(
-            Bracketed(
-                Ref("ExpressionSegment"),
-            ),
-            Ref("ExpressionSegment"),
-        ),
+        OptionallyBracketed(Ref("ExpressionSegment")),
         Dedent,
     )
 
@@ -1975,6 +1991,7 @@ class DropStatementSegment(BaseSegment):
         OneOf(
             "TABLE",
             "VIEW",
+            "USER",
         ),
         Ref("IfExistsGrammar", optional=True),
         Ref("TableReferenceSegment"),
@@ -2476,6 +2493,7 @@ class StatementSegment(BaseSegment):
         Ref("CreateFunctionStatementSegment"),
         Ref("CreateModelStatementSegment"),
         Ref("DropModelStatementSegment"),
+        Ref("DescribeStatementSegment"),
     )
 
     def get_table_references(self):
@@ -2505,4 +2523,21 @@ class WithNoSchemaBindingClauseSegment(BaseSegment):
         "NO",
         "SCHEMA",
         "BINDING",
+    )
+
+
+@ansi_dialect.segment()
+class DescribeStatementSegment(BaseSegment):
+    """A `Describe` statement.
+
+    DESCRIBE <object type> <object name>
+    """
+
+    type = "describe_statement"
+    match_grammar = StartsWith("DESCRIBE")
+
+    parse_grammar = Sequence(
+        "DESCRIBE",
+        Ref("NakedIdentifierSegment"),
+        Ref("ObjectReferenceSegment"),
     )