+++ src/sqlfluff/dialects/dialect_teradata.py
@@ -538,7 +538,7 @@ class FromUpdateClauseSegment(BaseSegment):
         "FROM",
         Delimited(
             # Optional old school delimited joins
-            Ref("TableExpressionSegment"),
+            Ref("FromExpressionElementSegment"),
             delimiter=Ref("CommaSegment"),
         ),
     )